@@ -39,7 +39,7 @@
     "prettier-plugin-svelte": "^3.1.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
     "prisma": "^5.5.2",
-    "svelte": "^4.0.5",
+    "svelte": "5.0.0-next.200",
     "svelte-check": "^3.4.3",
     "tailwindcss": "^3.3.5",
     "tslib": "^2.4.1",
