@@ -1,6 +1,5 @@
 <script lang="ts">
   import type { SignedInUser } from '$lib/shared/types';
-  import { createEventDispatcher } from 'svelte';
   import RoundLougout from 'virtual:icons/ic/round-logout';
   import RoundLogin from 'virtual:icons/ic/round-login';
   import RoundSearch from 'virtual:icons/ic/round-search';
@@ -13,7 +12,12 @@
   import NavigationElement from './NavigationElement.svelte';
   import { vibrate } from '$lib/actions/vibrate';
 
-  export let user: SignedInUser | null = null;
+  interface Props {
+    user?: SignedInUser | null;
+    onclickedelement?: () => void;
+  }
+
+  let { user = null, onclickedelement }: Props = $props();
 
   type NavigationElementType = {
     href: string;
@@ -69,14 +73,12 @@
       icon: RoundLogin
     }
   ];
-
-  const dispatch = createEventDispatcher();
 </script>
 
 {#if user}
   {#if user.role === 'ADMIN' || user.role === 'OWNER'}
     {#each adminElements as el}
-      <NavigationElement on:clickedelement {...el} />
+      <NavigationElement {onclickedelement} {...el} />
     {/each}
   {/if}
   {#each loggedInElements as el}
@@ -84,7 +86,7 @@
       <form class="px-16 sm:px-0" method="POST" action="/logout" use:enhance>
         <button
           class="flex w-full items-center justify-start gap-2 text-2xl sm:text-base"
-          on:click={() => dispatch('clickedelement')}
+          onclick={() => onclickedelement?.()}
           type="submit"
           use:vibrate
         >
@@ -93,11 +95,11 @@
         </button>
       </form>
     {:else}
-      <NavigationElement on:clickedelement {...el} />
+      <NavigationElement {onclickedelement} {...el} />
     {/if}
   {/each}
 {:else}
   {#each loggedOutElements as el}
-    <NavigationElement on:clickedelement {...el} />
+    <NavigationElement {onclickedelement} {...el} />
   {/each}
 {/if}
