@@ -2,11 +2,14 @@
   import { vibrate } from '$lib/actions/vibrate';
   import { getAlbumArtUrl } from '$lib/shared/fetchAlbumArt';
   import type { Release } from '$lib/shared/types';
-  import { createEventDispatcher } from 'svelte';
 
-  export let release: Release;
-  let showImage: boolean = false;
-  const dispatch = createEventDispatcher();
+  interface Props {
+    release: Release;
+    onchoose: (releaseId: string) => void;
+  }
+
+  let { release, onchoose }: Props = $props();
+  let showImage: boolean = $state(false);
 </script>
 
 <div class="flex w-40 flex-none flex-col items-center justify-center">
@@ -24,7 +27,7 @@
   <div class="m-2 flex overflow-clip rounded-md text-sm font-bold">
     <button
       class="bg-primary transition-all hover:bg-primary/80 disabled:cursor-not-allowed disabled:bg-primary/65 disabled:opacity-80 disabled:hover:bg-primary/65"
-      on:click={() => (showImage = true)}
+      onclick={() => (showImage = true)}
       disabled={showImage}
       use:vibrate={{ mute: showImage }}
     >
@@ -32,7 +35,7 @@
     </button>
     <button
       class="bg-primary transition-all hover:bg-primary/80"
-      on:click={() => dispatch('choosen', { releaseId: release.id })}
+      onclick={() => onchoose(release.id)}
       use:vibrate
     >
       Choose Image
