@@ -6,13 +6,19 @@
   import TrackRow from '$lib/components/TrackRow.svelte';
   import { enhance } from '$app/forms';
   import HeartOff from 'virtual:icons/iconamoon/heart-off';
-  import { playTrack } from '$lib/stores/audioPlayer.js';
   import { vibrate } from '$lib/actions/vibrate.js';
+  import { getAudioPlayer } from '$lib/states/audioPlayer.svelte.js';
+  import type { PageData } from './$types';
 
-  export let data;
+  interface Props {
+    data: PageData;
+  }
+
+  let { data }: Props = $props();
   const [send, receive] = crossfade;
-  let container: HTMLDivElement;
-  let scrolled = false;
+  let container: HTMLDivElement | null = $state(null);
+  let scrolled = $state(false);
+  const audioPlayer = getAudioPlayer();
 </script>
 
 <div class="flex h-full w-full flex-col">
@@ -43,28 +49,30 @@
   <div
     class="flex h-full flex-col overflow-auto"
     bind:this={container}
-    on:scroll={() => (scrolled = container?.scrollTop > 0)}
+    onscroll={() => (scrolled = !!container?.scrollTop && container?.scrollTop > 0)}
   >
     {#each data.playlist.tracks as track, index (track.id)}
       <div class="w-full flex-none" animate:flip={{ duration: 250 }}>
         <TrackRow
           handleClick={() => {
-            playTrack(data.playlist.tracks, index);
+            audioPlayer.playTrack(data.playlist.tracks, index);
           }}
           {track}
           delay={250 + index * 30}
         >
-          <form use:enhance slot="button" action="?/remove" method="POST">
-            <input type="hidden" name="trackId" value={track.id} />
-            <button
-              class="flex h-full w-full items-center justify-center"
-              title="Remove"
-              type="submit"
-              use:vibrate
-            >
-              <HeartOff class="text-2xl text-zinc-600 hover:text-primary" />
-            </button>
-          </form>
+          {#snippet button()}
+            <form use:enhance action="?/remove" method="POST">
+              <input type="hidden" name="trackId" value={track.id} />
+              <button
+                class="flex h-full w-full items-center justify-center"
+                title="Remove"
+                type="submit"
+                use:vibrate
+              >
+                <HeartOff class="text-2xl text-zinc-600 hover:text-primary" />
+              </button>
+            </form>
+          {/snippet}
         </TrackRow>
       </div>
     {:else}
